move the react benchmarks
